minor fixes

+++ ft_bzero.c
@@ -1,6 +1,6 @@
 #include "libft.h"
 
-void	ft_bzero(void *s, size_t n)
+void	ft_bzero(void *s, int n)
 {
 	ft_memset(s, 0, n);
 }
