+++ ft_memset.c
@@ -1,8 +1,8 @@
 #include "libft.h"
 
-void	*ft_memset(void *s, int c, size_t n)
+void	*ft_memset(void *s, int c, int n)
 {
-	size_t		i;
+	int		i;
 	char	*aux;
 
 	i = 0;
